Renamed contentProcessor to signatureProcessor
