+++ server/src/signatureProcessor.ts
